+++ azure-eventhubs-eph/src/main/java/com/microsoft/azure/eventprocessorhost/PartitionContext.java
@@ -37,6 +37,8 @@ public class PartitionContext
     {
         this.hostContext = hostContext;
         this.partitionId = partitionId;
+
+        this.runtimeInformation = new ReceiverRuntimeInformation(partitionId);
     }
 
     /***
