+++ azure-eventhubs/src/main/java/com/microsoft/azure/eventhubs/ReceiverDisconnectedException.java
@@ -17,6 +17,7 @@ package com.microsoft.azure.eventhubs;
  * @see <a href="http://go.microsoft.com/fwlink/?LinkId=761101">http://go.microsoft.com/fwlink/?LinkId=761101</a>
  */
 public class ReceiverDisconnectedException extends EventHubException {
+
     private static final long serialVersionUID = 3385140843418138213L;
 
     ReceiverDisconnectedException() {
