+++ azure-eventhubs/src/main/java/com/microsoft/azure/eventhubs/RetryPolicy.java
@@ -12,6 +12,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 // TODO: SIMPLIFY retryPolicy - ConcurrentHashMap is not needed
 public abstract class RetryPolicy {
+
     private static final RetryPolicy NO_RETRY = new RetryExponential(Duration.ofSeconds(0), Duration.ofSeconds(0), 0, ClientConstants.NO_RETRY);
 
     private final String name;
