+++ azure-eventhubs/src/main/java/com/microsoft/azure/eventhubs/ServerBusyException.java
@@ -11,6 +11,7 @@ package com.microsoft.azure.eventhubs;
  * @see <a href="http://go.microsoft.com/fwlink/?LinkId=761101">http://go.microsoft.com/fwlink/?LinkId=761101</a>
  */
 public class ServerBusyException extends EventHubException {
+
     private static final long serialVersionUID = 1L;
 
     /**
